+++ test/plug/accesslog/formatter_test.exs
@@ -6,7 +6,9 @@ defmodule Plug.AccessLog.FormatterTest do
 
   test "no format means default format" do
     datetime = :calendar.local_time()
-    conn     = conn(:get, "/") |> put_private(:plug_accesslog, datetime)
+    conn     =
+         conn(:get, "/")
+      |> put_private(:plug_accesslog, %{ local_time: datetime })
 
     assert Formatter.format(nil, conn) == Formatter.format(:default, conn)
   end