+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Additional formatting directives
   - `%a` - Remote IP-address
+  - `%T` - Time taken to serve the request (full seconds)
 
 ## v0.6.0 (2015-05-23)